requirements.txt
bkcharts==0.2
bokeh==0.12.6
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
matplotlib==2.0.2
numpy==1.13.1
pillow==4.2.1
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.3
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
tornado==4.5.1
Werkzeug==0.12.2
